Sorted RecyclerView items
Implemented sort into the RecyclerView item list by submitting the list of database items after they have been sorted chronologically and then in terms of if they have passed during the current year or not.

For example, if the current date is March 20th, and two birthdays have been saved into the database, one on January 15th, and the other on March 25th, the RecyclerView will show the birthday on March 25th first, and then the birthday in January. This is in order to provide an better user experience.