Added notification feature
Added a UniquePeriodicWorker that aims to execute at noon daily and will send a notification to the user when the date for a birthday has come around. I also tweaked the CalendarFragment such that it displays whatever birthday is today as at the top of the screen.

The UniquePeriodicWorker will be discarded and replaced with a new UPW when CalendarFragment's view is created, however. This means that another notification will be sent:
 1. Every time that the app is launched.
 2. Every time a user goes to add or edit a birthday and comes back to CalendarFragment.

 This may be annoying for some users. A fix is in progress.