Modified notification intent & pendingIntent
Modified notification Intent and PendingIntents, such that the app does not launch when the notification is pressed. The (Pending)Intent was not useful, as it led the user back to the app where they would not be provided with any new information, so this was a improvement over the past version of the app. The issue with the repeated notifications mentioned in the previous commit has still not been resolved.